requirements.txt
streamlit
openai-agents
python-dotenv
google-generativeai
nest_asyncio
